requirements.txt
anthropic>=0.40.0
streamlit>=1.32.0
python-docx>=1.1.0
duckdb>=0.10.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
